+++ ssllabs/__init__.py
@@ -1,6 +1,7 @@
 """Qualys SSL Labs API in Python."""
 from importlib.metadata import PackageNotFoundError, version
 
+from .exceptions import SsllabsOverloadedError, SsllabsUnavailableError
 from .ssllabs import Ssllabs
 from .trust_store import TrustStore
 
@@ -12,4 +13,4 @@ except PackageNotFoundError:
     # package is not installed
     __version__ = "0.0.0"
 
-__all__ = ["Ssllabs", "TrustStore", "__license__", "__version__"]
+__all__ = ["Ssllabs", "SsllabsOverloadedError", "SsllabsUnavailableError", "TrustStore", "__license__", "__version__"]