+++ ssllabs/api/endpoint.py
@@ -27,9 +27,9 @@ class Endpoint(_Api):
                             is intended for API consumers that don't want to wait for assessment results. Can't be used at the
                             same time as the startNew parameter.
         :type fromCache: str
-        :raises httpx.ConnectTimeout: SSL Labs Servers don't respond.
-        :raises httpx.HTTPStatusError: A client or server error response occurred.
-        :raises httpx.ReadTimeout: SSL Labs Servers don't respond.
+        :raises SsllabsUnavailableError: The SSL Labs service is not available.
+        :raises SsllabsOverloadedError: The SSL Labs service is overloaded. You should reduce your usage or wait a bit.
+        :raises HTTPStatusError: Something unexpected happened. Please file us a bug.
         """
         self._verify_kwargs(kwargs.keys(), ["fromCache"])
         r = await self._call("getEndpointData", host=host, s=s, **kwargs)