+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 
 - Drop support for Python 3.7
+- Exceptions have been reworked and are now mostly independent from httpx
 
 ## [v1.2.1] - 2023/08/07